Remove CommunicationChannel, create CommandManager
* Move `CommunicationChannel::Input::Command` functions to `CommandManager` class.

* Remove comIn.cpp and comOut.cpp.

* Rename comCommmands.cpp to command.cpp.

* Create command.h.

* Replace `com.out.throwError()` with a cout operator<<  function.

* Move `operator<< ` template (for ints, chars, doubles, etc) to comT.h.
