+++ MinePlusPlus/comT.h
@@ -2,5 +2,10 @@
 #define comT_h
 
 #include "com.h"
+template <typename T>
+OStream& operator<< (OStream& out, T input) {
+  Serial.print(input);
+  return out;
+}
 
 #endif
